+++ CHANGELOG.rst
@@ -45,6 +45,7 @@ Version 0.10 (unreleased)
 * Addition of a ``segmentize`` function for GEOS >= 3.10 (#299)
 * Addition of ``oriented_envelope`` and ``minimum_rotated_rectangle`` functions (#314)
 * Addition of ``minimum_bounding_circle`` and ``minimum_bounding_radius`` functions for GEOS >= 3.8 (#315)
+* Addition of a ``shortest_line`` ("nearest points") function (#334)
 
 **Bug fixes**